Move testUserCanDeleteFreestyleProject from the DeleteFreestyleProjectTest class to the _FreestyleTest class and remove the DeleteFreestyleProjectTest class
